+++ pencil/include/pencil_layouter.h
@@ -276,10 +276,10 @@ static inline void pencil_layouter_is_pos_on_grid ( const pencil_layouter_t *thi
  *  \param out_y_count number of y-position grid lines
  */
 static inline void pencil_layouter_get_grid_lines ( const pencil_layouter_t *this_,
-                                                    int32_t *out_x0,
-                                                    int32_t *out_y0,
-                                                    int32_t *out_dx,
-                                                    int32_t *out_dy,
+                                                    double *out_x0,
+                                                    double *out_y0,
+                                                    double *out_dx,
+                                                    double *out_dy,
                                                     uint32_t *out_x_count,
                                                     uint32_t *out_y_count
                                                   );
